+++ test/io/postgrest.py
@@ -230,3 +230,20 @@ def is_ipv6(addr):
         return True
     except OSError:
         return False
+
+
+def set_statement_timeout(postgrest, role, milliseconds):
+    """Set the statement timeout for the given role.
+    For this to work reliably with low previous timeout settings,
+    use a postgrest instance that doesn't use the affected role."""
+
+    response = postgrest.session.post(
+        "/rpc/set_statement_timeout", data={"role": role, "milliseconds": milliseconds}
+    )
+    assert response.text == ""
+    assert response.status_code == 204
+
+
+def reset_statement_timeout(postgrest, role):
+    "Reset the statement timeout for the given role to the default 0 (no timeout)"
+    set_statement_timeout(postgrest, role, 0)
